translation/requirements.txt
torch==1.9.0
pytorch-lightning==1.4.9
tensorboardX==2.4
test-tube==0.7.5
sacrebleu==2.0.0
adapter-transformers==2.1.2